Refactor issuesubmitter.py -> issue_submitter.py
